note eval
